fix(cli): enforce sane ranges on reassembly threshold flags (#96)
## Summary
Closes the three process-gap deferrals the 3-pass adversarial review of
the snaplen / small-segment sub-cycle logged in STATE.md (drift items
6–8).

- **Drift 6** — the reassembly threshold flags accepted any value, so an
operator could silently configure a nonsensical detector
(`--small-segment-max-bytes 100000` makes *every* TCP segment "small").
The flags now carry clap `value_parser` ranges enforced at parse time,
matching the Snort ranges the config field docs already cite:
`--overlap-threshold` 0–255, `--small-segment-threshold` 0–2048,
`--small-segment-max-bytes` 0–2048. `small_segment_max_bytes` is now
`u16`. `--out-of-window-threshold` stays unbounded — no NIDS prior art
gives it a defensible ceiling.
- **Drift 7** — the threshold-boundary test pattern existed only for the
small-segment detector. Added exactly-threshold negative tests for the
overlap and out-of-window detectors, so a `>` → `>=` regression is
caught for all three.
- **Drift 8** — documented (in `Cargo.toml` and `src/decoder.rs`) that
the decoder's strict→lax `SliceError::Len` discrimination is coupled to
the etherparse 0.16 API. `"0.16"` already excludes 0.17 via Cargo's
caret semantics; the IPv6 / corrupt-packet tests act as contract tests
for a future bump.

The Snort range values (overlap_limit 0–255, small_segments 0–2048) were
established by the earlier P2.05 research dispatches and are already
cited in the config field docs — this PR enforces what the docs already
document.

## Test plan
- [x] `cargo fmt --check`
- [x] `cargo clippy --all-targets -- -D warnings`
- [x] `cargo test --all-targets` — all green. New: out-of-range
threshold flag values are rejected at parse time; overlap and
out-of-window exactly-threshold runs stay silent.

+++ src/reassembly/config.rs
@@ -84,9 +84,9 @@ pub struct ReassemblyConfig {
     /// conservative default — small enough to flag deliberate
     /// fragmentation, large enough to also catch an attacker who splits
     /// into 8–15 byte segments rather than literal 1-byte ones. Setting
-    /// it to `0` disables small-segment detection entirely. Tune via
-    /// `--small-segment-max-bytes`.
-    pub small_segment_max_bytes: usize,
+    /// it to `0` disables small-segment detection entirely. The
+    /// `--small-segment-max-bytes` flag enforces Snort's 0–2048 range.
+    pub small_segment_max_bytes: u16,
     /// Ports on which small-segment detection is suppressed. A flow is
     /// exempt when **either** endpoint port appears here — a small-segment
     /// run on these ports is benign interactive traffic, not evasion.
